switch and write CSV file

+++ src/main/java/com/example/banksimulation/CDAccount.java
@@ -17,7 +17,7 @@ public class CDAccount extends Account {
 
 
 
-    CDAccount(int accountNumber, int accountBalance, Customer accountHolder) {
+    CDAccount(int accountNumber, double accountBalance, Customer accountHolder) {
         super(accountNumber, accountBalance, accountHolder);
     }
